backend/requirements.txt
 
google-generativeai
fastapi
uvicorn[standard]
pandas
numpy
faiss-cpu
cohere
python-dotenv
pydantic
python-multipart
sqlalchemy
psycopg2-binary
requests